pyproject.toml
[project]
name = "newsagg"
version = "0.6.0"
description = "A multi-source news aggregator with a REST API"
authors = [
    { name="Igor Paim", email="igorpaim8@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.1.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6",
    "aiohttp>=3.9",
    "httpx>=0.27.0",
    "feedparser>=6.0",
    "sqlalchemy>=2.0.41",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.21.0",
    "python-jose>=3.5.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.20",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff",               # Linter
    "mypy",               # Type checker
    "ipython",            # Better REPL
]

[project.urls]
"Homepage" = "https://github.com/igrorp"
"Repository" = "https://github.com/igrorp/newsagg"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 80
